requirements.txt
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi==0.61.2
h11==0.11.0
idna==2.10
lxml==4.6.1
pydantic==1.7.2
requests==2.24.0
soupsieve==2.0.1
starlette==0.13.6
urllib3==1.25.11
uvicorn==0.12.2
python-dotenv==0.15.0
selenium==3.141.0